external/PNDM/requirements.txt
torch>=1.7.0
tensorboard
pytorch-fid
PyYAML
#mpi4py
scipy
numpy
lmdb
tqdm
